requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
transformers
sentence-transformers
faiss-cpu
opencv-python
Pillow
torch
streamlit